requirements.txt
numpy
cloudpickle==2.0.*
paramiko==2.*
scp
tqdm
redis
pyyaml
humanfriendly
pydantic
